backend/python/requirements.txt
Flask
mysql-connector-python
llama-cpp-python
python-dotenv
SQLAlchemy
flask-cors